Add spark ETL container to wizard docker-compose
When setting up the docker-compose used to start up the wizard stack
(elasticsearch, kibabna, dashboard-importer), now the spark ETL job
is also included in the mix:

A `docker-compose up` will now also trigger the ETL job.

The ETL job waits for elasticsearch to be available before attempting
to import data to it.

In order for gerrit analytics endpoint to be available to the
dockerized ETL job, the host ip address and gerrit port need
to be configured in the docker compose:

- The host ip address is read using InetAddress object
- The gerrit schema and port are read from configuration

Feature: Issue 9867
Change-Id: I3419a635d8d7333e4d2ea078daec3e19287cbac0