Show waiting bar modal when loading dashboard
When docker-compose up command is issued via 'analytics-wizard~server'
endpoint, the page keeps polling 'analytics-wizard~status' endpoint
to check the status of the import.

The polling continues until 202 is received (job is still in progress).
When 204 is received the user is redirected to the kibana dashboard.
Any other polling result is shown as error message.

Additionally, the analytics-wizard~server has been modified in order
to run the docker-compose in detached mode, to avoid the http request
hanging forever.

Feature: Issue 9870
Change-Id: I4116db8ae454e489522d97452f5eeb06f6c9cc6b